added guiinterface

+++ src/glbrick/TriangleSpec.java
@@ -62,7 +62,7 @@ public class TriangleSpec implements BrickSpec {
 	
 	public DrawnObject toDrawnObject() 
 	{
-		float[] temp = new float[]{1f,1f,1f}; 
+		double[] temp = new double[]{1,1,1}; 
 		return new DrawnObject(vertices, temp);
 	}
 }
